Break out bootloader state machine to prepare for unit-testing
